test(e2e): implement review comments
- Removed getEnforcementOnExistingPodsTest because it's not possible
anymore.
- Moved utility functions to helpers_test.go
- Flipped the argument order of go test command
